Actualización Proyecto GestionImagenFranciscoAMP.
- Se modifican los archivos actualizar, finActualizar, leerActualizar, insertar, leerInsertar e index.jsp
- Se renombran carpetas delete y read denominándolas eliminar y visualizar.
- Dentro de la carpeta eliminar se crean, de nuevo los archivos eliminar, leerEliminar y finEliminar.
- Se renombra el archivo read, denominándose visualizar.
- es.albarregas.beans, dentro del cual se actualiza Usuario.java cambiando el tipo de dato de int a byte.

04/06/2020
by Francisco Antonio Murillo Pacheco

+++ src/main/java/es/albarregas/beans/Usuario.java
@@ -8,16 +8,16 @@ import java.sql.Date;
  * @author Framcisco_Antonio
  */
 public class Usuario {
-    private int id;
+    private Byte id;
     private String nombre;
     private Date fechaNacimiento;
     private String avatar;
 
-    public int getId() {
+    public Byte getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Byte id) {
         this.id = id;
     }
